@@ -4,7 +4,7 @@ import 'package:arcore_flutter_plugin/src/arcore_augmented_image.dart';
 import 'package:arcore_flutter_plugin/src/arcore_rotating_node.dart';
 import 'package:arcore_flutter_plugin/src/utils/vector_utils.dart';
 import 'package:flutter/services.dart';
-import 'package:meta/meta.dart';
+// import 'package:meta/meta.dart';
 import 'arcore_hit_test_result.dart';
 
 import 'arcore_node.dart';
@@ -32,11 +32,19 @@ class ArCoreController {
     return arcoreInstalled;
   }
 
+  final bool enableUpdateListener;
+  final bool enableTapRecognizer;
+  final bool enablePlaneRenderer;
+  final bool debug;
+  late MethodChannel _channel;
+  late StringResultHandler onError;
+  late StringResultHandler onNodeTap;
+
   ArCoreController(
-      {int id,
-      this.enableTapRecognizer,
-      this.enablePlaneRenderer,
-      this.enableUpdateListener,
+      {required int id,
+      required this.enableTapRecognizer,
+      required this.enablePlaneRenderer,
+      required this.enableUpdateListener,
       this.debug = false
 //    @required this.onUnsupported,
       }) {
@@ -45,19 +53,11 @@ class ArCoreController {
     init();
   }
 
-  final bool enableUpdateListener;
-  final bool enableTapRecognizer;
-  final bool enablePlaneRenderer;
-  final bool debug;
-  MethodChannel _channel;
-  StringResultHandler onError;
-  StringResultHandler onNodeTap;
-
 //  UnsupportedHandler onUnsupported;
-  ArCoreHitResultHandler onPlaneTap;
-  ArCorePlaneHandler onPlaneDetected;
+  late ArCoreHitResultHandler onPlaneTap;
+  late ArCorePlaneHandler onPlaneDetected;
   String trackingState = '';
-  ArCoreAugmentedImageTrackingHandler onTrackingImage;
+  late ArCoreAugmentedImageTrackingHandler onTrackingImage;
 
   init() async {
     try {
@@ -134,7 +134,7 @@ class ArCoreController {
     return Future.value();
   }
 
-  Future<void> addArCoreNode(ArCoreNode node, {String parentNodeName}) {
+  Future<void> addArCoreNode(ArCoreNode node, {String? parentNodeName}) {
     assert(node != null);
     final params = _addParentNodeNameToParams(node.toMap(), parentNodeName);
     if (debug) {
@@ -144,17 +144,16 @@ class ArCoreController {
     return _channel.invokeMethod('addArCoreNode', params);
   }
 
-
-  Future<String> togglePlaneRenderer() async {
+  Future<dynamic> togglePlaneRenderer() async {
     return _channel.invokeMethod('togglePlaneRenderer');
   }
 
-  Future<String> getTrackingState() async {
+  Future<dynamic> getTrackingState() async {
     return _channel.invokeMethod('getTrackingState');
   }
 
   addArCoreNodeToAugmentedImage(ArCoreNode node, int index,
-      {String parentNodeName}) {
+      {String? parentNodeName}) {
     assert(node != null);
 
     final params = _addParentNodeNameToParams(node.toMap(), parentNodeName);
@@ -163,7 +162,7 @@ class ArCoreController {
   }
 
   Future<void> addArCoreNodeWithAnchor(ArCoreNode node,
-      {String parentNodeName}) {
+      {String? parentNodeName}) {
     assert(node != null);
     final params = _addParentNodeNameToParams(node.toMap(), parentNodeName);
     if (debug) {
@@ -176,13 +175,13 @@ class ArCoreController {
     return _channel.invokeMethod('addArCoreNodeWithAnchor', params);
   }
 
-  Future<void> removeNode({@required String nodeName}) {
+  Future<void> removeNode({required String nodeName}) {
     assert(nodeName != null);
     return _channel.invokeMethod('removeARCoreNode', {'nodeName': nodeName});
   }
 
   Map<String, dynamic> _addParentNodeNameToParams(
-      Map geometryMap, String parentNodeName) {
+      Map<String, dynamic> geometryMap, String? parentNodeName) {
     if (parentNodeName?.isNotEmpty ?? false)
       geometryMap['parentNodeName'] = parentNodeName;
     return geometryMap;
@@ -199,7 +198,7 @@ class ArCoreController {
 
   void _handlePositionChanged(ArCoreNode node) {
     _channel.invokeMethod<void>('positionChanged',
-        _getHandlerParams(node, convertVector3ToMap(node.position.value)));
+        _getHandlerParams(node, convertVector3ToMap(node.position.value)!));
   }
 
   void _handleRotationChanged(ArCoreRotatingNode node) {
@@ -209,7 +208,7 @@ class ArCoreController {
 
   void _updateMaterials(ArCoreNode node) {
     _channel.invokeMethod<void>(
-        'updateMaterials', _getHandlerParams(node, node.shape.toMap()));
+        'updateMaterials', _getHandlerParams(node, node.shape!.toMap()));
   }
 
   Map<String, dynamic> _getHandlerParams(
@@ -219,7 +218,7 @@ class ArCoreController {
     return values;
   }
 
-  Future<void> loadSingleAugmentedImage({@required Uint8List bytes}) {
+  Future<void> loadSingleAugmentedImage({required Uint8List bytes}) {
     assert(bytes != null);
     return _channel.invokeMethod('load_single_image_on_db', {
       'bytes': bytes,
@@ -227,14 +226,14 @@ class ArCoreController {
   }
 
   Future<void> loadMultipleAugmentedImage(
-      {@required Map<String, Uint8List> bytesMap}) {
+      {required Map<String, Uint8List> bytesMap}) {
     assert(bytesMap != null);
     return _channel.invokeMethod('load_multiple_images_on_db', {
       'bytesMap': bytesMap,
     });
   }
 
-  Future<void> loadAugmentedImagesDatabase({@required Uint8List bytes}) {
+  Future<void> loadAugmentedImagesDatabase({required Uint8List bytes}) {
     assert(bytes != null);
     return _channel.invokeMethod('load_augmented_images_database', {
       'bytes': bytes,
@@ -242,11 +241,11 @@ class ArCoreController {
   }
 
   void dispose() {
-    _channel?.invokeMethod<void>('dispose');
+    _channel.invokeMethod<void>('dispose');
   }
 
   void resume() {
-    _channel?.invokeMethod<void>('resume');
+    _channel.invokeMethod<void>('resume');
   }
 
   Future<void> removeNodeWithIndex(int index) async {
