@@ -13,8 +13,8 @@ class ArCoreFaceView extends StatefulWidget {
   final bool debug;
 
   const ArCoreFaceView(
-      {Key key,
-      this.onArCoreViewCreated,
+      {required Key key,
+      required this.onArCoreViewCreated,
       this.enableAugmentedFaces = false,
       this.debug = false})
       : super(key: key);
@@ -27,7 +27,7 @@ class _ArCoreFaceViewState extends State<ArCoreFaceView>
     with WidgetsBindingObserver {
   @override
   void initState() {
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance?.addObserver(this);
     super.initState();
   }
 
@@ -63,7 +63,7 @@ class _ArCoreFaceViewState extends State<ArCoreFaceView>
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance?.removeObserver(this);
     super.dispose();
   }
 }
