@@ -5,14 +5,14 @@ import 'package:vector_math/vector_math_64.dart';
 import 'package:arcore_flutter_plugin/src/shape/arcore_shape.dart';
 
 class ArCoreCube extends ArCoreShape {
+  final Vector3 size;
+
   ArCoreCube({
-    this.size,
-    List<ArCoreMaterial> materials,
+    required this.size,
+    required List<ArCoreMaterial> materials,
   }) : super(
           materials: materials,
         );
-
-  final Vector3 size;
 
   @override
   Map<String, dynamic> toMap() => <String, dynamic>{
